Refactor View and Form into components
