+++ src/pages/CruiseDetail.jsx
@@ -3,22 +3,18 @@ import React, { useState } from "react";
 import {
   Button,
   Grid,
-  Fieldset,
   Tag,
-  Form,
-  Label,
-  TextInput,
-  DatePicker,
-  Select
 } from "@trussworks/react-uswds";
 import StationSummary from "../components/StationSummary";
 import StationNew from "../components/StationNew";
-import DescriptionListItem from "../components/DescriptionListItem";
+import HeaderWithEdit from "../components/HeaderWithEdit";
+import CruiseView from "../components/CruiseView";
+import CruiseForm from "../components/CruiseForm";
+import AppCard from "../components/AppCard";
 import { useNavigate, useParams } from "react-router-dom";
-import { listValueLookup, CruiseStatus } from "../utils/listLookup";
+import { listValueLookup } from "../utils/listLookup";
 import { setStatusColor } from "../utils/setStatusColor";
 import { generateTzDateTime, getLocationTz } from "../utils/dateTimeHelpers";
-import { post, put } from "../utils/requestMethods";
 import { usePortsList, useCruiseStatusesList } from "../hooks/useListTables";
 import { useGetCruiseById, useGetStationsByCruiseId, useUpdateCruise, useAddStation } from "../hooks/useCruises";
 
@@ -70,8 +66,6 @@ const CruiseDetailPage = () => {
     returnPortId,
   } = cruise;
   const cruiseStatus = listValueLookup(cruiseStatuses, cruiseStatusId);
-  const departurePort = listValueLookup(ports, departurePortId);
-  const returnPort = listValueLookup(ports, returnPortId);
 
   const handleNavCruisesList = () => {
     navigate("/cruises");
@@ -152,137 +146,20 @@ const CruiseDetailPage = () => {
           <Tag className={`padding-1 usa-tag--big ${setStatusColor(cruiseStatusId)}`}>{cruiseStatus}</Tag>
         </div>
       </Grid>
-      <div className="border radius-lg padding-1 margin-y-2 app-card">
-        <Grid row className="flex-justify-end margin-bottom-1">
-          {
-            activeAction === CruiseAction.EDIT
-              ? <Button
-                className="margin-right-0"
-                onClick={() => setActiveAction(null)}
-                secondary
-              >
-                Cancel Edit Cruise
-              </Button>
-              : <Button
-                className="margin-right-0"
-                onClick={() => setActiveAction(CruiseAction.EDIT)}
-                disabled={activeAction !== null && activeAction !== CruiseAction.EDIT}
-              >
-                Edit Cruise
-              </Button>
-          }
-
-        </Grid>
+      <AppCard>
+        <HeaderWithEdit
+          title=""
+          eventType={CruiseAction.EDIT}
+          activeAction={activeAction}
+          handleSetAction={() => setActiveAction(CruiseAction.EDIT)}
+          handleCancelAction={() => setActiveAction(null)} />
         {activeAction !== null && activeAction === CruiseAction.EDIT
           ?
-          <Form className="maxw-full" onSubmit={handleSaveCruise}>
-            <Grid row gap>
-              <Grid col={12} tablet={{ col: true }}>
-                <Grid row>
-                  <Label htmlFor="cruise-name" className="grid-col-4 text-bold margin-top-2" requiredMarker>
-                    Cruise Name:
-                  </Label>
-                  <TextInput id="cruise-name" name="cruiseName" className="grid-col-8" defaultValue={cruiseName} required />
-                </Grid>
-              </Grid>
-              <Grid col={12} tablet={{ col: true }}>
-                <Grid row>
-                  <Label htmlFor="vessel-name" className="grid-col-4 text-bold margin-top-2" requiredMarker>
-                    Vessel Name:
-                  </Label>
-                  <TextInput id="vessel-name" name="vesselName" className="grid-col-8" defaultValue={vesselName} required />
-                </Grid>
-              </Grid>
-            </Grid>
-            <Grid row gap>
-              <Grid col={12} tablet={{ col: true }}>
-                <Fieldset legend="Departure" className="app-legend-bold-text">
-                  <Grid row>
-                    <Label htmlFor="start-date" className="grid-col-4 text-bold margin-top-2" requiredMarker>
-                      Date:
-                    </Label>
-                    <DatePicker id="start-date" name="startDate" className="grid-col-8 margin-top-0" defaultValue={startDate} required />
-                  </Grid>
-                  <Grid row>
-                    <Label htmlFor="departure-port-select" className="grid-col-4 text-bold margin-top-2" requiredMarker>
-                      Port:
-                    </Label>
-                    <Select
-                      id="departure-port-select"
-                      name="departurePortId"
-                      className="grid-col-8"
-                      defaultValue={departurePortId}
-                      required
-                    >
-                      <option value={null}>- Select Port -</option>
-                      {ports.map((port) => (
-                        <option key={port.id} value={port.id}>
-                          {port.name}
-                        </option>
-                      ))}
-                    </Select>
-                  </Grid>
-                </Fieldset>
-              </Grid>
-              <Grid col>
-                <Fieldset legend="Return" className="app-legend-bold-text">
-                  <Grid row>
-                    <Label htmlFor="end-date" className="grid-col-4 text-bold margin-top-2" >
-                      Date:
-                    </Label>
-                    <DatePicker id="end-date" name="endDate" className="grid-col-8 margin-top-0" defaultValue={endDate} />
-                  </Grid>
-                  <Grid row>
-                    <Label htmlFor="return-port-select" className="grid-col-4 text-bold margin-top-2" >
-                      Port:
-                    </Label>
-                    <Select
-                      id="return-port-select"
-                      name="returnPortId"
-                      className="grid-col-8"
-                      defaultValue={returnPortId}
-                    >
-                      <option value={null}>- Select Port -</option>
-                      {ports.map((port) => (
-                        <option key={port.id} value={port.id}>
-                          {port.name}
-                        </option>
-                      ))}
-                    </Select>
-                  </Grid>
-                </Fieldset>
-              </Grid>
-            </Grid>
-            <Grid row className="flex-column flex-align-end">
-              <Button type="submit" className="margin-right-0">Save Cruise</Button>
-            </Grid>
-          </Form>
-          : <>
-            <Grid row gap>
-              <Grid col={12} tablet={{ col: true }}>
-                <DescriptionListItem term="Cruise Name:" description={cruiseName} />
-              </Grid>
-              <Grid col={12} tablet={{ col: true }}>
-                <DescriptionListItem term="Vessel Name:" description={vesselName} />
-              </Grid>
-            </Grid>
-            <Grid row className="margin-top-2">
-              <Grid col={12} tablet={{ col: true }}>
-                <Fieldset legend="Departure" className="app-legend-bold-text">
-                  <DescriptionListItem term="Date:" description={startDate} />
-                  <DescriptionListItem term="Port:" description={departurePort} />
-                </Fieldset>
-              </Grid>
-              <Grid col>
-                <Fieldset legend="Return" className="app-legend-bold-text">
-                  <DescriptionListItem term="Date:" description={endDate} />
-                  <DescriptionListItem term="Port:" description={returnPort} />
-                </Fieldset>
-              </Grid>
-            </Grid>
-          </>
+          <CruiseForm cruise={cruise} ports={ports} handleSubmit={handleSaveCruise} />
+          :
+          <CruiseView cruise={cruise} ports={ports} />
         }
-      </div>
+      </AppCard>
       <Grid row className="flex-justify margin-bottom-1">
         <h2 className="app-sec-header">Stations</h2>
         {
